Move hooks into proper folder
