+++ src/components/Apps/AppsDevModeHome.tsx
@@ -22,7 +22,7 @@ import { Add } from '@mui/icons-material';
 import { QORTAL_APP_CONTEXT, getBaseApiReact } from '../../App';
 import { executeEvent } from '../../utils/events';
 import { Spacer } from '../../common/Spacer';
-import { useModal } from '../../common/useModal';
+import { useModal } from '../../hooks/useModal.tsx';
 import { createEndpoint, isUsingLocal } from '../../background/background.ts';
 import { Label } from '../Group/AddGroup';
 import ShortUniqueId from 'short-unique-id';
